Add Visual Studio requirement for Windows in the README
This commit highlights the need for Windows users to have Visual Studio
including the "Desktop development with C++" workload already installed
before trying to get gen-subs on their computer.
Otherwise, when that requirement is not met, you get an error to
node-gyp requirements on Windows.
See https://github.com/nodejs/node-gyp#on-windows